Refactor: 각각의 페이지 폴더에 index.js 추가해 폴더 경로 수정
